% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tika.R
\name{tika_parse_to_html}
\alias{tika_parse_to_html}
\title{Apache tika parser to html in R}
\usage{
tika_parse_to_html(file)
}
\arguments{
\item{file}{file path}
}
\value{
html
}
\description{
Apache tika parser to html in R
}
\examples{
\dontrun{
tika_parse_to_html("presentation.pptx")
}
}
