% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tika.R
\name{tika_parse_to_xml}
\alias{tika_parse_to_xml}
\title{Apache tika parser to xml in R}
\usage{
tika_parse_to_xml(file)
}
\arguments{
\item{file}{file path}
}
\value{
xml
}
\description{
Apache tika parser to xml in R
}
\examples{
tika_parse_to_xml("presentation.pptx")

}
